- Enhancement for #99: Private Key File can now also be read from memory. - Refactor of NewSftp and NewSSH
