+++ PoshSSH/PoshSSH/PoshSSH.csproj
@@ -51,6 +51,7 @@
     <Compile Include="GetScpFile.cs" />
     <Compile Include="GetScpFolder.cs" />
     <Compile Include="GetSftpFile.cs" />
+    <Compile Include="NewSessionBase.cs" />
     <Compile Include="SetSftpFile.cs" />
     <Compile Include="SshModHelper.cs" />
     <Compile Include="NewSftpSession.cs" />
